Use while loop to complete assignment logic.
The assignment given wanted the student to reach the 5 lines answer by running a code with a while loop. The answer should be 5 lines, with stars (*) on each, that gradually adds more stars as the lines go; To simplify, the first line, for example, should have 1 star, the second line 2 stars, and so on until we reach the required number of stars in the last line (in our case, 5).